feat: adicionado melhores gestão de segurança / doc: documentação de codigo simples
