--- Arkanum-api/src/main/java/io/github/cavalcante_dev/Arcanum/controller/dto/CreateUserDTO.java
@@ -1,4 +1,0 @@
-package io.github.cavalcante_dev.Arcanum.controller.dto;
-
-public record CreateUserDTO(String username, String password) {
-}
