+++ Arkanum-api/src/main/java/io/github/cavalcante_dev/Arkanum/controller/UserController.java
@@ -1,4 +1,4 @@
-package io.github.cavalcante_dev.Arcanum.controller;
+package io.github.cavalcante_dev.Arkanum.controller;
 
 import java.util.List;
 import java.util.Set;
@@ -13,12 +13,14 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
-import io.github.cavalcante_dev.Arcanum.controller.dto.CreateUserDTO;
-import io.github.cavalcante_dev.Arcanum.entitys.Role;
-import io.github.cavalcante_dev.Arcanum.entitys.User;
-import io.github.cavalcante_dev.Arcanum.repository.RoleRepository;
-import io.github.cavalcante_dev.Arcanum.repository.UserRepository;
+import io.github.cavalcante_dev.Arkanum.controller.dto.CreateUserDTO;
+import io.github.cavalcante_dev.Arkanum.entitys.Role;
+import io.github.cavalcante_dev.Arkanum.entitys.User;
+import io.github.cavalcante_dev.Arkanum.repository.RoleRepository;
+import io.github.cavalcante_dev.Arkanum.repository.UserRepository;
 import jakarta.transaction.Transactional;
+
+// Controlador de funções do usuário, tudo relacionado a edição e cadastro estará aqui.
 
 @RestController
 public class UserController {
@@ -26,6 +28,8 @@ public class UserController {
     private final UserRepository userRepository;
     private final RoleRepository roleRepository;
     private final BCryptPasswordEncoder passwordEncoder;
+
+    // Construtor de Classe
 
     public UserController(UserRepository userRepository,
                           RoleRepository roleRepository,
@@ -35,27 +39,39 @@ public class UserController {
         this.passwordEncoder = passwordEncoder;
     }
 
+    // Função de cadastro de usuário no sistema.
+
     @Transactional
     @PostMapping("/cadastro")
     public ResponseEntity<Void> newUser(@RequestBody CreateUserDTO dto) {
 
         var basicRole = roleRepository.findByName(Role.Values.BASIC.name());
 
+        if (basicRole == null) {
+            basicRole = new Role();
+            basicRole.setName(Role.Values.BASIC.name());
+            roleRepository.save(basicRole);
+        }
+
         var userFromDb = userRepository.findByUsername(dto.username());
+
         if (userFromDb.isPresent()){
-            System.out.println("Usuario já cadastrado");
+            System.out.println("Usuário já cadastrado");
             throw new ResponseStatusException(HttpStatus.UNPROCESSABLE_ENTITY);
         }
 
         var user = new User();
         user.setUsername(dto.username());
         user.setPassword(passwordEncoder.encode(dto.password()));
+        user.setName(dto.name());
         user.setRole(Set.of(basicRole));
 
         userRepository.save(user);
 
         return ResponseEntity.ok().build();
     }
+
+    // Função de listar todos os usuários cadastrados, usados apenas por admins.
 
     @GetMapping("/users")
     @PreAuthorize("hasAuthority('SCOPE_admin')")
