+++ Arkanum-api/src/main/java/io/github/cavalcante_dev/Arkanum/controller/dto/CreateUserDTO.java
@@ -1,0 +1,4 @@
+package io.github.cavalcante_dev.Arkanum.controller.dto;
+
+public record CreateUserDTO(String username, String name, String password) {
+}
