+++ Arkanum-api/src/main/java/io/github/cavalcante_dev/Arkanum/entitys/Role.java
@@ -1,0 +1,55 @@
+package io.github.cavalcante_dev.Arkanum.entitys;
+
+import jakarta.persistence.*;
+
+// Entidade que determina a função do usuário, entre ADMIN e BASIC
+
+@Entity
+@Table(name = "tb_roles")
+public class Role {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "role_id")
+    private Long roleId;
+
+    private String name;
+
+    // Denomina os valores de ADMIN e BASIC para funcionalidades exclusivas
+
+    public enum Values {
+
+        ADMIN(1),
+        BASIC(2);
+
+        long roleId;
+
+        Values(long roleId) {
+            this.roleId = roleId;
+        }
+
+        public long getRoleId() {
+            return roleId;
+        }
+
+    }
+
+    //Getters and Setters
+
+    public Long getRoleId() {
+        return roleId;
+    }
+
+    public void setRoleId(Long roleId) {
+        this.roleId = roleId;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+}
