+++ Arkanum-api/src/main/java/io/github/cavalcante_dev/Arkanum/entitys/User.java
@@ -1,23 +1,34 @@
-package io.github.cavalcante_dev.Arcanum.entitys;
+package io.github.cavalcante_dev.Arkanum.entitys;
 
-import io.github.cavalcante_dev.Arcanum.controller.dto.LoginRequest;
+import io.github.cavalcante_dev.Arkanum.controller.dto.LoginRequest;
 import jakarta.persistence.*;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.util.Set;
 import java.util.UUID;
 
+//Entidade - usuário: utilizado para login e cadastro é a entidade principal.
+
 @Entity
 @Table(name = "tb_users")
 public class User {
+
+    // Gera um ID automaticamente
 
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     @Column(name = "user_id")
     private UUID userID;
+
+    // Faz com que não existam usernames repetidos
+
     @Column(unique = true)
     private String username;
+
+    private String name;
     private String password;
+
+    //Dá para o usuário a função de usuário ou admin
 
     @ManyToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER)
     @JoinTable(
@@ -26,6 +37,16 @@ public class User {
             inverseJoinColumns = @JoinColumn(name = "role_id")
     )
     private Set<Role> role;
+
+    //Getters and Setters
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
 
     public UUID getUserID() {
         return userID;
@@ -58,6 +79,8 @@ public class User {
     public void setRole(Set<Role> role) {
         this.role = role;
     }
+
+    // Verifica se o login do usuário está correto.
 
     public boolean isLoginCorrect(LoginRequest loginRequest, PasswordEncoder passwordEncoder) {
         return passwordEncoder.matches(loginRequest.password(), this.password);
