+++ Arkanum-api/src/main/java/io/github/cavalcante_dev/Arkanum/repository/CharacterSheetRepository.java
@@ -1,7 +1,6 @@
-package io.github.cavalcante_dev.Arcanum.repository;
+package io.github.cavalcante_dev.Arkanum.repository;
 
-import io.github.cavalcante_dev.Arcanum.entitys.CharacterSheet;
-import io.github.cavalcante_dev.Arcanum.entitys.Role;
+import io.github.cavalcante_dev.Arkanum.entitys.CharacterSheet;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
