+++ Arkanum-api/src/main/java/io/github/cavalcante_dev/Arkanum/repository/RoleRepository.java
@@ -1,11 +1,8 @@
-package io.github.cavalcante_dev.Arcanum.repository;
+package io.github.cavalcante_dev.Arkanum.repository;
 
-import io.github.cavalcante_dev.Arcanum.entitys.Role;
-import io.github.cavalcante_dev.Arcanum.entitys.User;
+import io.github.cavalcante_dev.Arkanum.entitys.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
 
 @Repository
 public interface RoleRepository extends JpaRepository<Role, Long> {
