--- Arkanum-api/src/main/resources/app.key
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDPUp/84GcQirNb
-hFdiLXR57xq2qOdUGDPOqSC5Cr4Y6NU5X5/URGlJHUomqK77Yg0xvI/cCK/CsuQj
-z6oT9KK/l/A5zI8vQdLdumS7SzMVmuMrLA6OCW/hobNOvejqaSZohMuU0dvI9ize
-5txHRFdyl2LANMLVfPs+LkjUMqUb78F1YZu4aZ1NnRw0aQDI78m8h3+RInCF4QhH
-VCrHW3CzQ1lwfjHfLxkWPo4al+smNgXLtWlb8cZUYNo1wyVEzk+mfeJJyUE1UTdK
-cVq3T83303OcGZP7rq6jXZoXGS4iyzxRODfqakEVNcX+YfEH7h81cwXntXOJ4IZa
-gZukZw/FAgMBAAECggEAHaQ4j5Ud5zWNczY2sCjDkHNpJFh9yn4zWgbAUPSCPVl8
-stlDJrwpOXTglypi6joYnNXS06jCSFn/WEdmj7fYVQBKLlla12zX0y/FGyojQaGF
-fTStWA1tW5sc4dMxK+KgxSzHVp2yiMb5n1zufZn9zrpqOgTRHmjiGv/PFdev15Br
-/aI8WRvo+Gf3FslIdKyZ5YrhGiNxPs784uial/wSB3DfyvkFse3GJ6OMmB3/FVQD
-9DgdwwOAjAfvmuxLjTGgnRqtbpVgWt5L8W33yVzahZUSi4sHy82CRC8cNlD69Cil
-f+0GddDGru8ABLHrCJy1wxAJtkHR/+jp00BeiQyTuQKBgQDn2qju15q3XCvXo7H1
-fSEL39bY9BOux968GQbX/DztvpEfq8lo4QsGnUnBH7XVL1UGqGbw/dNd/XQPEeVA
-t6pY40HzO3LYNebyk1a5zH0ueMHIejnfv009OdQEwH8UenOsDzPlI7rOYu+qSLG6
-ezvT1ajCuSlIhzzQ91rt35AprwKBgQDk6fJg1W99g0H2FAK+okcaeiYBWMtYRn/f
-0+H27GDY3f6xAKBwnCB2pe9ifa5ceLvp/CA+QHfQjjXATCQYa8a/xbZfecwVm0rT
-qt3VOnlXI0fezqo6JE1jKLGgHleeykkzrYn34fkJrnUhIK4W5alSpdjirNbNeEKB
-vuP3VWSeywKBgCxx4x5d4D+0AIeGzq9jfDk9gH74UumXUOA33/L5XGiqtkEe+i4L
-rcRn1GKcBIxIVYFAOPOLhMAovN3CrXK5goMllVfeFobR1mK70c0uIPUOmv8mRv/Y
-OSM6eaUrjXDMhG4AQA91yPKVTqIS61b3FqHl/ELPodVQb10Tfa/JwmXrAoGBAKvJ
-nWlqAzcv76paLg0d/zoxfJ/6IO/oSEKS4brKfYuCXYY7sCuSHxKrSFCjfNxMhsj7
-ur685q2Oa2waN+RTzV1RZKGV58A3QBpE2iEQtsrb4S2l7j2fzKujeFlvK9YwfGrZ
-yvZJkPjhAWmQ+tSYGU+6wl6p0rFzlIWF2dJdNRutAoGAe2hDtTMPZfdxz9ygF5NF
-6SN0UnCeaRHP0WZiPdPTheR88GvLF+RRjQNqdtAb7soMNDDcpHe+gyzelpzmf06I
-KWHw0UYhHr8w8Zbtk3xB9C5ieW+dkZr8eiTC6nL4Dya5umKfMuwKwjD6cK038WyY
-ZbYXxOmQkU1j62myNtNjVsI=
------END PRIVATE KEY-----
