--- Arkanum-api/src/main/resources/app.pub
@@ -1,9 +1,0 @@
------BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAz1Kf/OBnEIqzW4RXYi10
-ee8atqjnVBgzzqkguQq+GOjVOV+f1ERpSR1KJqiu+2INMbyP3AivwrLkI8+qE/Si
-v5fwOcyPL0HS3bpku0szFZrjKywOjglv4aGzTr3o6mkmaITLlNHbyPYs3ubcR0RX
-cpdiwDTC1Xz7Pi5I1DKlG+/BdWGbuGmdTZ0cNGkAyO/JvId/kSJwheEIR1Qqx1tw
-s0NZcH4x3y8ZFj6OGpfrJjYFy7VpW/HGVGDaNcMlRM5Ppn3iSclBNVE3SnFat0/N
-99NznBmT+66uo12aFxkuIss8UTg36mpBFTXF/mHxB+4fNXMF57VzieCGWoGbpGcP
-xQIDAQAB
------END PUBLIC KEY-----
