+++ Arkanum-api/src/main/resources/application.properties
@@ -1,14 +1,19 @@
 spring.application.name=Arcanum
 
-jwt.public.key=classpath:app.pub
-jwt.private.key=classpath:app.key
-
 logging.level.org.springframework.security=TRACE
 logging.level.org.springframework.web=DEBUG
 
+spring.config.import=application.secret.properties
+
+jwt.public.key=${PUBLIC_KEY}
+jwt.private.key=${PRIVATE_KEY}
+
+app.admin.password=${ADMIN_SENHA}
+app.admin.user=${ADMIN_USER}
+
 spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/arcanum
-spring.datasource.username=myuser
-spring.datasource.password=2506
+spring.datasource.url=${app.db.path}
+spring.datasource.username=${app.db.user}
+spring.datasource.password=${app.db.password}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.show-sql: true
