+++ Arkanum-api/src/main/resources/application.secret.properties
@@ -1,0 +1,9 @@
+app.db.path=${DB_PATH}
+app.db.user=${DB_USER}
+app.db.password=${DB_PASSWORD}
+
+PRIVATE_KEY=classpath:keys/app.key
+PUBLIC_KEY=classpath:keys/app.pub
+
+ADMIN_SENHA=2569
+ADMIN_USER=ARKANUMCODEX
