+++ Arkanum-api/src/main/resources/static/key.env
@@ -1,0 +1,8 @@
+
+PRIVATE_KEY=MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDPUp/84GcQirNbhFdiLXR57xq2qOdUGDPOqSC5Cr4Y6NU5X5/URGlJHUomqK77Yg0xvI/cCK/CsuQjz6oT9KK/l/A5zI8vQdLdumS7SzMVmuMrLA6OCW/hobNOvejqaSZohMuU0dvI9ize5txHRFdyl2LANMLVfPs+LkjUMqUb78F1YZu4aZ1NnRw0aQDI78m8h3+RInCF4QhHVCrHW3CzQ1lwfjHfLxkWPo4al+smNgXLtWlb8cZUYNo1wyVEzk+mfeJJyUE1UTdKcVq3T83303OcGZP7rq6jXZoXGS4iyzxRODfqakEVNcX+YfEH7h81cwXntXOJ4IZagZukZw/FAgMBAAECggEAHaQ4j5Ud5zWNczY2sCjDkHNpJFh9yn4zWgbAUPSCPVl8stlDJrwpOXTglypi6joYnNXS06jCSFn/WEdmj7fYVQBKLlla12zX0y/FGyojQaGFfTStWA1tW5sc4dMxK+KgxSzHVp2yiMb5n1zufZn9zrpqOgTRHmjiGv/PFdev15Br/aI8WRvo+Gf3FslIdKyZ5YrhGiNxPs784uial/wSB3DfyvkFse3GJ6OMmB3/FVQD9DgdwwOAjAfvmuxLjTGgnRqtbpVgWt5L8W33yVzahZUSi4sHy82CRC8cNlD69Cilf+0GddDGru8ABLHrCJy1wxAJtkHR/+jp00BeiQyTuQKBgQDn2qju15q3XCvXo7H1fSEL39bY9BOux968GQbX/DztvpEfq8lo4QsGnUnBH7XVL1UGqGbw/dNd/XQPEeVAt6pY40HzO3LYNebyk1a5zH0ueMHIejnfv009OdQEwH8UenOsDzPlI7rOYu+qSLG6ezvT1ajCuSlIhzzQ91rt35AprwKBgQDk6fJg1W99g0H2FAK+okcaeiYBWMtYRn/f0+H27GDY3f6xAKBwnCB2pe9ifa5ceLvp/CA+QHfQjjXATCQYa8a/xbZfecwVm0rTqt3VOnlXI0fezqo6JE1jKLGgHleeykkzrYn34fkJrnUhIK4W5alSpdjirNbNeEKBvuP3VWSeywKBgCxx4x5d4D+0AIeGzq9jfDk9gH74UumXUOA33/L5XGiqtkEe+i4LrcRn1GKcBIxIVYFAOPOLhMAovN3CrXK5goMllVfeFobR1mK70c0uIPUOmv8mRv/YOSM6eaUrjXDMhG4AQA91yPKVTqIS61b3FqHl/ELPodVQb10Tfa/JwmXrAoGBAKvJnWlqAzcv76paLg0d/zoxfJ/6IO/oSEKS4brKfYuCXYY7sCuSHxKrSFCjfNxMhsj7ur685q2Oa2waN+RTzV1RZKGV58A3QBpE2iEQtsrb4S2l7j2fzKujeFlvK9YwfGrZyvZJkPjhAWmQ+tSYGU+6wl6p0rFzlIWF2dJdNRutAoGAe2hDtTMPZfdxz9ygF5NF6SN0UnCeaRHP0WZiPdPTheR88GvLF+RRjQNqdtAb7soMNDDcpHe+gyzelpzmf06IKWHw0UYhHr8w8Zbtk3xB9C5ieW+dkZr8eiTC6nL4Dya5umKfMuwKwjD6cK038WyYZbYXxOmQkU1j62myNtNjVsI=
+PUBLIC_KEY=MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAz1Kf/OBnEIqzW4RXYi10ee8atqjnVBgzzqkguQq+GOjVOV+f1ERpSR1KJqiu+2INMbyP3AivwrLkI8+qE/Siv5fwOcyPL0HS3bpku0szFZrjKywOjglv4aGzTr3o6mkmaITLlNHbyPYs3ubcR0RXcpdiwDTC1Xz7Pi5I1DKlG+/BdWGbuGmdTZ0cNGkAyO/JvId/kSJwheEIR1Qqx1tws0NZcH4x3y8ZFj6OGpfrJjYFy7VpW/HGVGDaNcMlRM5Ppn3iSclBNVE3SnFat0/N99NznBmT+66uo12aFxkuIss8UTg36mpBFTXF/mHxB+4fNXMF57VzieCGWoGbpGcPxQIDAQAB
+API_DATABASE=2506
+DB_PATH=jdbc:mysql://localhost:3306/arcanum
+DB_USER=myuser
+ADMIN_SENHA=2569
+ADMIN_USER=ARKANUMCODEX
