+++ Arkanum-api/src/test/java/io/github/cavalcante_dev/Arkanum/ArcanumApplicationTests.java
@@ -1,4 +1,4 @@
-package io.github.cavalcante_dev.Arcanum;
+package io.github.cavalcante_dev.Arkanum;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
